Move ExperimentConfig to its own file
